refactor inner class to own file
